Robot/requirements.txt
websockets>=10.0
pyserial>=3.5